Refactoring function locations in daemon and common
