+++ common/git_test.go
@@ -1,0 +1,24 @@
+package common
+
+import (
+	"os"
+	"path"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCheckForGit(t *testing.T) {
+	cwd, _ := os.Getwd()
+	assert.NotEqual(t, nil, CheckForGit(cwd))
+	inertia, _ := path.Split(cwd)
+	assert.Equal(t, nil, CheckForGit(inertia))
+}
+
+func TestGetSSHRemoteURL(t *testing.T) {
+	httpsURL := "https://github.com/ubclaunchpad/inertia.git"
+	sshURL := "git@github.com:ubclaunchpad/inertia.git"
+
+	assert.Equal(t, sshURL, GetSSHRemoteURL(httpsURL))
+	assert.Equal(t, sshURL, GetSSHRemoteURL(sshURL))
+}
