+++ common/util_test.go
@@ -19,35 +19,6 @@ var (
 	testToken      = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30.AqFWnFeY9B8jj7-l3z0a9iaZdwIca7xhUF3fuaJjU90"
 )
 
-func TestGenerateToken(t *testing.T) {
-	token, err := GenerateToken(testPrivateKey)
-	assert.Nil(t, err, "generateToken must not fail")
-	assert.Equal(t, token, testToken)
-}
-
-func TestGetGithubKey(t *testing.T) {
-	inertiaKeyPath := path.Join(os.Getenv("GOPATH"), "/src/github.com/ubclaunchpad/inertia/test_env/test_key")
-	pemFile, err := os.Open(inertiaKeyPath)
-	assert.Nil(t, err)
-	_, err = GetGithubKey(pemFile)
-	assert.Nil(t, err)
-}
-
-func TestGetSSHRemoteURL(t *testing.T) {
-	httpsURL := "https://github.com/ubclaunchpad/inertia.git"
-	sshURL := "git@github.com:ubclaunchpad/inertia.git"
-
-	assert.Equal(t, sshURL, GetSSHRemoteURL(httpsURL))
-	assert.Equal(t, sshURL, GetSSHRemoteURL(sshURL))
-}
-
-func TestCheckForGit(t *testing.T) {
-	cwd, _ := os.Getwd()
-	assert.NotEqual(t, nil, CheckForGit(cwd))
-	inertia, _ := path.Split(cwd)
-	assert.Equal(t, nil, CheckForGit(inertia))
-}
-
 func TestCheckForDockerCompose(t *testing.T) {
 	cwd, err := os.Getwd()
 	assert.Nil(t, err)
