+++ daemon.go
@@ -6,7 +6,6 @@ import (
 
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"github.com/ubclaunchpad/inertia/common"
 	"github.com/ubclaunchpad/inertia/daemon"
 )
 
@@ -50,7 +49,7 @@ var tokenCmd = &cobra.Command{
 			log.Fatal(err)
 		}
 
-		token, err := common.GenerateToken(keyBytes.([]byte))
+		token, err := daemon.GenerateToken(keyBytes.([]byte))
 		if err != nil {
 			log.Fatal(err)
 		}
