+++ daemon/auth_test.go
@@ -1,0 +1,13 @@
+package daemon
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGenerateToken(t *testing.T) {
+	token, err := GenerateToken(testPrivateKey)
+	assert.Nil(t, err, "generateToken must not fail")
+	assert.Equal(t, token, testToken)
+}
